refactor: extract portable app controller
