+++ README.md
@@ -2,7 +2,9 @@
 
 **Your Hermes, carried forward.**
 
-Celeste is a native Android client for a self-hosted [Hermes Agent](https://github.com/NousResearch/hermes-agent) dashboard. It connects directly to the same Hermes server as Desktop, so your profiles, conversations, and agent work stay together.
+Celeste is an Android-first native client for a self-hosted [Hermes Agent](https://github.com/NousResearch/hermes-agent) dashboard. It connects directly to the same Hermes server as Desktop, so your profiles, conversations, and agent work stay together.
+
+Android is the only application target today. Celeste is being structured as one Compose Multiplatform product so a future iOS target can share protocol behavior, application state, and custom UI instead of becoming a separately maintained client.
 
 The installed Android app is named **Hermes Celeste** so it is easy to find in the launcher and app list.
 
@@ -17,7 +19,7 @@ The installed Android app is named **Hermes Celeste** so it is easy to find in t
 
 ## Status
 
-Celeste is in early development. The native chat flow is functional; OAuth sign-in, persistent credentials, rich Markdown, attachments, approvals, and broader Hermes management features are still ahead.
+Celeste is in early development. The native Android chat flow and secure connection restoration are functional; OAuth sign-in, rich Markdown, attachments, approvals, and broader Hermes management features are still ahead. There is no iOS application or release commitment yet.
 
 ## Development
